Move Options component to Common folder
